# src/main/webapp/META-INF/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.13
  title: Outcarcerate
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /attorney:
    post:
      tags:
        - attorney
      summary: Add a new attorney
      description: ''
      operationId: addAttorney
      responses:
        '405':
          description: Invalid input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attorney'
        required: true
    get:
      tags:
        - attorney
      summary: Get a list of attorneys
      description: Returns a all attorneys
      operationId: getAllAttorneys
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttorneyResponse'
  '/attorney/{attorneyId}':
    get:
      tags:
        - attorney
      summary: Find attorney by ID
      description: Returns a single attorney
      operationId: getAttorneyById
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttorneyResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Attorney not found
    delete:
      tags:
        - attorney
      summary: Delete attorney by ID
      description: Delete a single attorney
      operationId: deleteAttorneyById
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Attorney not found
...
  /case:
    post:
      tags:
        - case
      summary: Add a new case to an attorney
      description: ''
      operationId: addCase
      parameters:
        - in: query
          name: attorneyId
          required: true
          description: Enter the attorney id for which the case to be added
          schema:
            type: integer
      responses:
        '405':
          description: Invalid input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Case'
        required: true
  '/case/{attorneyId}':
    get:
      tags:
        - case
      summary: Get all cases for attorney
      description: Returns a single case
      operationId: getCaseById
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Case not found
  '/case/{attorneyId}/{caseId}':
    get:
      tags:
        - case
      summary: Get all cases for attorney
      description: Returns a single case
      operationId: getCase
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney
          required: true
          schema:
            type: string
        - name: caseId
          in: path
          description: ID of case
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Case not found
    delete:
      tags:
        - case
      summary: Delete case by ID
      description: Delete a single case
      operationId: deleteCaseById
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney
          required: true
          schema:
            type: string
        - name: caseId
          in: path
          description: ID of case to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Case not found
  '/case/report/{attorneyId}/{caseId}':
    get:
      tags:
        - case
      summary: Get report for the case
      description: Returns a single case
      operationId: getCaseReport
      parameters:
        - name: attorneyId
          in: path
          description: ID of attorney
          required: true
          schema:
            type: string
        - name: caseId
          in: path
          description: ID of case
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseReport'
        '400':
          description: Invalid ID supplied
        '404':
          description: Case not found
servers:
  - url: /v1
components:
  schemas:
    Charge:
      type: object
      required:
        - charge_code
        - attempted
        - primary
      properties:
        _id:
          type: string
        _rev:
          type: string
        trial_type:
          type: string
          enum:
            - Guilty plea
            - Trial by judge
            - Trial by jury
        disposition_charged_class:
          type: string
        charge_code:
          type: string
          enum:
            - Administration of Justice
            - Aggravated Assault Police Officer
            - Aggravated Battery Police Officer
            - Aggravated Battery With A Firearm
            - Aggravated Battery
            - Aggravated Discharge Firearm
            - Aggravated DUI
            - Aggravated Fleeing and Eluding
            - Aggravated Identity Theft
            - Aggravated Robbery
            - Antitrust
            - Armed Robbery
            - Armed Violence
            - Arson and Attempt Arson
            - Arson
            - Attempt Armed Robbery
            - Attempt Arson
            - Attempt First Degree Murder
            - Attempt Homicide
            - Attempt Sex Crimes
            - Attempt Vehicular Hijacking
            - Battery
            - Bomb Threat
            - Bribery
            - Bribery/Corruption
            - Burglary
            - Burglary/Trespass
            - Child Abduction
            - Child Pornography
            - Commercialized Vice
            - Communicating With Witness
            - Credit Card Cases
            - Criminal Damage to Property
            - Criminal Trespass To Residence
            - Deceptive Practice
            - Disarming Police Officer
            - Dog Fighting
            - Domestic Battery
            - Driving With Suspended Or Revoked License
            - Drug Possession
            - Drug Trafficking
            - DUI
            - Environmental
            - Escape - Failure to Return
            - Extortion/Racketeering
            - FALSIFICATION OF ACCOUNTS
            - Failure to Register as a Sex Offender
            - Firearms
            - Food and Drug
            - Forgery
            - Forgery/Counter/Copyright
            - Fraud
            - Fraud/Theft/Embezzlement
            - Fraudulent ID
            - Gambling
            - Gun - Non UUW
            - Gun Running
            - Hate Crimes
            - Home Invasion
            - Homicide
            - Human Trafficking
            - Identity Theft
            - Immigration
            - Impersonating Police Officer
            - Intimidation
            - Invidual Rights
            - Kidnapping
            - Major Accidents
            - Manslaughter
            - Money Laundering
            - Murder
            - Narcotics
            - National Defense
            - Obscenity/Other Sex Offenses
            - Obstructing Justice
            - Official Misconduct
            - Other Offense
            - Other
            - Pandering
            - Perjury
            - Possession Of Burglary Tools
            - Possession of Contraband in Penal Institution
            - Possession of Explosives
            - Possession of Stolen Motor Vehicle
            - Prison Offenses
            - PROMIS Conversion
            - Prostitution
            - Reckless Discharge of Firearm
            - Reckless Homicide
            - Residential Burglary
            - Retail Theft
            - REVOKED/SUSPENDED 2ND+ DUI
            - Robbery
            - Sex Crimes
            - SEX WITH ANIMAL/<18 PRESENT
            - Sexual Abuse
            - Stalking
            - Stalking/Harassing
            - Tampering
            - Tax
            - Theft by Deception
            - Theft
            - Unlawful Restraint
            - UUW - Unlawful Use of Weapon
            - Vehicular Hijacking
            - Vehicular Invasion
            - Violate Bail Bond
            - VIO BAIL BOND/CLASS X CONVIC
            - Violation of Sex Offender Registration
            - Violation Order Of Protection
        attempted:
          type: boolean
        primary:
          type: boolean
        controlled_substance_quantity_level:
          type: integer
          minimum: 0
          maximum: 17
        possible_sentences:
          type: array
          items:
            $ref: '#/components/schemas/Sentence'
    Sentence:
      type: object
      required:
        - sentence_type
      properties:
        _id:
          type: string
        _rev:
          type: string
        chargeDisposition:
          type: string
          enum:
            - BFW
            - Death Suggested-Cause Abated
            - Finding Guilty
            - FNG
            - FNPC
            - Nolle Prosecution
            - 'Null'
            - Plea of Guilty
            - SOL
            - Verdict-Not Guilty
            - Verdict Guilty
        phase:
          type: string
        sentence_type:
          type: string
          enum:
            - Prison Only
            - Prison and Alternatives
            - Probation Only
            - Probation and Alternatives
            - Fine Ordered/No Restitution
            - Restitution Ordered/No Fine
            - Both Fine & Restitution Ordered
            - Prison
            - Supervision
            - Conditional Discharge
            - Probation
            - Jail
            - Conversion
            - Cook County Boot Camp
            - Probation Terminated Satisfactorily
            - Inpatient Mental Health Services
            - Probation Terminated Unsatisfactorily
            - Conditional Release
            - Probation Terminated Instanter
            - Death
            - 2nd Chance Probation
        commitmentTerm:
          type: integer
          format: int64
        commitmentUnit:
          type: string
        commitmentType:
          type: string
        minimum_incarceration_months:
          type: integer
          format: int64
        maximum_incarceration_months:
          type: integer
          format: int64
        probation_term_months:
          type: integer
          format: int64
        minimum_probation_months:
          type: integer
          format: int64
        maximum_probation_months:
          type: integer
          format: int64
        fine_dollars:
          type: integer
          format: int64
        minimum_fine_dollars:
          type: integer
          format: int64
        maximum_fine_dollars:
          type: integer
          format: int64
        community_service_hours:
          type: integer
          format: int64
        restitution:
          type: string
        alternative_sentence:
          type: string
        suspended_sentence:
          type: boolean
    Case:
      type: object
      required:
        - possible_charges
      properties:
        _id:
          type: string
        _rev:
          type: string
        client_id:
          type: string
        attorney_id:
          type: string
        possible_charges:
          type: array
          items:
            $ref: '#/components/schemas/Charge'
...
    Deviation:
      type: object
      required:
        - charge_code
      properties:
        _id:
          type: string
        _rev:
          type: string
        charge_code:
          type: string
        sentence_deviations:
          type: array
          items:
            $ref: '#/components/schemas/Sentence'
    Attorney:
      type: object
      required:
        - username
      properties:
        _id:
          type: string
        _rev:
          type: string
        username:
          type: string
        cases:
          type: array
          items:
            $ref: '#/components/schemas/Case'
...
    AttorneyResponse:
      type: object
      required:
        - success
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        error:
          type: string
        warning:
          type: string
        attorney:
          type: array
          items:
            $ref: '#/components/schemas/Attorney'
...
    BiasDetectionResponse:
      type: object
      required:
        - success
        - charge_code
        - race
        - gender
        - controlled_substance_quantity_level
        - deviations
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        error:
          type: string
        warning:
          type: string
        charge_code:
          type: string
          enum:
            - Administration of Justice
            - Aggravated Assault Police Officer
            - Aggravated Battery Police Officer
            - Aggravated Battery With A Firearm
            - Aggravated Battery
            - Aggravated Discharge Firearm
            - Aggravated DUI
            - Aggravated Fleeing and Eluding
            - Aggravated Identity Theft
            - Aggravated Robbery
            - Antitrust
            - Armed Robbery
            - Armed Violence
            - Arson and Attempt Arson
            - Arson
            - Attempt Armed Robbery
            - Attempt Arson
            - Attempt First Degree Murder
            - Attempt Homicide
            - Attempt Sex Crimes
            - Attempt Vehicular Hijacking
            - Battery
            - Bomb Threat
            - Bribery
            - Bribery/Corruption
            - Burglary
            - Burglary/Trespass
            - Child Abduction
            - Child Pornography
            - Commercialized Vice
            - Communicating With Witness
            - Credit Card Cases
            - Criminal Damage to Property
            - Criminal Trespass To Residence
            - Deceptive Practice
            - Disarming Police Officer
            - Dog Fighting
            - Domestic Battery
            - Driving With Suspended Or Revoked License
            - Drug Possession
            - Drug Trafficking
            - DUI
            - Environmental
            - Escape - Failure to Return
            - Extortion/Racketeering
            - FALSIFICATION OF ACCOUNTS
            - Failure to Register as a Sex Offender
            - Firearms
            - Food and Drug
            - Forgery
            - Forgery/Counter/Copyright
            - Fraud
            - Fraud/Theft/Embezzlement
            - Fraudulent ID
            - Gambling
            - Gun - Non UUW
            - Gun Running
            - Hate Crimes
            - Home Invasion
            - Homicide
            - Human Trafficking
            - Identity Theft
            - Immigration
            - Impersonating Police Officer
            - Intimidation
            - Invidual Rights
            - Kidnapping
            - Major Accidents
            - Manslaughter
            - Money Laundering
            - Murder
            - Narcotics
            - National Defense
            - Obscenity/Other Sex Offenses
            - Obstructing Justice
            - Official Misconduct
            - Other Offense
            - Other
            - Pandering
            - Perjury
            - Possession Of Burglary Tools
            - Possession of Contraband in Penal Institution
            - Possession of Explosives
            - Possession of Stolen Motor Vehicle
            - Prison Offenses
            - PROMIS Conversion
            - Prostitution
            - Reckless Discharge of Firearm
            - Reckless Homicide
            - Residential Burglary
            - Retail Theft
            - REVOKED/SUSPENDED 2ND+ DUI
            - Robbery
            - Sex Crimes
            - SEX WITH ANIMAL/<18 PRESENT
            - Sexual Abuse
            - Stalking
            - Stalking/Harassing
            - Tampering
            - Tax
            - Theft by Deception
            - Theft
            - Unlawful Restraint
            - UUW - Unlawful Use of Weapon
            - Vehicular Hijacking
            - Vehicular Invasion
            - Violate Bail Bond
            - VIO BAIL BOND/CLASS X CONVIC
            - Violation of Sex Offender Registration
            - Violation Order Of Protection
        race:
          type: string
          enum:
            - White
            - Black
            - Hispanic
            - Other
            - Asian
            - American Indian
            - Biracial
            - 'White [Hispanic or Latino]'
            - 'White/Black [Hispanic or Latino]'
            - Unknown
        gender:
          enum:
            - Female
            - Male
            - Other
            - Unknown
          type: string
        controlled_substance_quantity_level:
          type: integer
          minimum: 0
          maximum: 17
        deviations:
          type: array
          items:
            $ref: '#/components/schemas/Deviation'
...
    CaseResponse:
      type: object
      required:
        - success
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        error:
          type: string
        warning:
          type: string
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Case'
    CaseReport:
      type: object
      required:
        - success
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        error:
          type: string
        bias:
          $ref: '#/components/schemas/BiasDetectionResponse'
        discrepancy:
          $ref: '#/components/schemas/Discrepancy'
    Discrepancy:
      type: object
      required:
        - model_name
      properties:
        model_name:
          type: string
        sentencing_discrepancy:
          type: number
          format: float
        severity:
          type: number
          format: float